--- webapp_front/dataconnect-generated/js/default-connector/esm/index.esm.js
@@ -1,7 +1,0 @@
-
-export const connectorConfig = {
-  connector: 'default',
-  service: 'webappfront',
-  location: 'us-central1'
-};
-
