--- webapp_front/dataconnect-generated/js/default-connector/index.cjs.js
@@ -1,7 +1,0 @@
-
-const connectorConfig = {
-  connector: 'default',
-  service: 'webappfront',
-  location: 'us-central1'
-};
-exports.connectorConfig = connectorConfig;
