--- webapp_front/dataconnect-generated/js/default-connector/index.d.ts
@@ -1,10 +1,0 @@
-import { ConnectorConfig } from 'firebase/data-connect';
-
-export const connectorConfig: ConnectorConfig;
-
-export type TimestampString = string;
-export type UUIDString = string;
-export type Int64String = string;
-export type DateString = string;
-
-
