--- webapp_front/dataconnect-generated/js/default-connector/react/esm/package.json
@@ -1,1 +1,0 @@
-{"type":"module"}
