--- webapp_front/dataconnect-generated/js/default-connector/react/index.d.ts
@@ -1,6 +1,0 @@
-import {  } from '../';
-import { } from '@tanstack-query-firebase/react/data-connect';
-import { } from '@tanstack/react-query';
-import { DataConnect } from 'firebase/data-connect';
-import { FirebaseError } from 'firebase/app';
-
