--- webapp_front/dataconnect-generated/js/default-connector/react/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "@firebasegen/default-connector-react",
-  "version": "1.0.0",
-  "author": "Firebase <firebase-support@google.com> (https://firebase.google.com/)",
-  "description": "Generated SDK For default",
-  "license": "Apache-2.0",
-  "engines": {
-    "node": " >=18.0"
-  },
-  "typings": "index.d.ts",
-  "main": "index.cjs.js",
-  "module": "esm/index.esm.js",
-  "browser": "esm/index.esm.js",
-  "peerDependencies": {
-    "@tanstack-query-firebase/react": "^2.0.0"
-  }
-}
