--- webapp_front/dataconnect/connector/connector.yaml
@@ -1,7 +1,0 @@
-connectorId: default
-generate:
-  javascriptSdk:
-    outputDir: ../../dataconnect-generated/js/default-connector
-    package: "@firebasegen/default-connector"
-    packageJsonDir: ../..
-    react: true
