--- webapp_front/dataconnect/dataconnect.yaml
@@ -1,12 +1,0 @@
-specVersion: "v1"
-serviceId: "webappfront"
-location: "us-central1"
-schema:
-  source: "./schema"
-  datasource:
-    postgresql:
-      database: "fdcdb"
-      cloudSql:
-        instanceId: "webappfront-fdc"
-      # schemaValidation: "COMPATIBLE"
-connectorDirs: ["./connector"]
