+++ webapp_front/src/App.css
@@ -23,3 +23,11 @@
   left: 50%;
   transform: translate(-50%, -50%);
 }
+
+.navs {
+  transition: 0.3s ease;
+  
+  &:hover {
+    color: var(--color-accent);
+  }
+}
